Fixed SwiftLint error
